No type uint defined in C.
It needs to be defined somewhere (usually as a typedef to unsigned int).
Anyway here we could just use guint which is already appropriately
defined (as the error indicates well), or simply int. I go for int,
seeing how it is used to loop through an enum, that is just all right.
Fixes:
> error: unknown type name 'uint'; did you mean 'guint'?