Keep stemmer_inc_dirs empty per default.
It is very unwise to set any kind of hardcoded path in a build system.
Unfortunately here it seems indeed that libstemmer does not have a .pc
file nor any other kind of prefix discovery system.
Probably in such case, the best will be to add an option (such as
"-Dlibstemmer_path=/some/thing") to allow overriding the library path.

In any case, for cross-compilation, we just cannot leave a hardcoded
path (to the common Linux system) around. As a first change, at least, I
make so this include path is not set to anything when not compiling with
stemming option.